pyproject.toml
[project]
name = "incident-response-rag"
version = "0.1.0"
description = "Incident Response RAG Pipeline - Processes incident response documents from Google Drive"
readme = "README.md"
requires-python = ">=3.11, <3.14"

dependencies = [
    "kfp>=2.10.0",
    "google-cloud-aiplatform>=1.116.0",
    "google-auth>=2.0.0",
    "google-api-python-client>=2.0.0",
    "pymupdf>=1.23.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "Pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "langchain-text-splitters>=1.0.0",
    "google-cloud-bigquery>=3.0",
    "google-cloud-discoveryengine>=0.11.0",
    "google-cloud-storage>=2.10.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "ruff>=0.1.0",
]

[project.optional-dependencies]
lint = [
    "ruff>=0.4.6",
    "mypy~=1.15.0",
    "codespell~=2.2.0",
    "types-PyYAML>=6.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "C",  # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP", # pyupgrade
    "RUF", # ruff specific rules
]
ignore = ["E501", "C901", "B006"] # ignore line too long, too complex

[tool.ruff.lint.isort]
known-first-party = ["pipeline"]

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = false
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return", "attr-defined"]

exclude = [".venv", "test_*.py", "tests"]

[[tool.mypy.overrides]]
module = [
    "google.cloud.storage",
    "google.cloud.discoveryengine",
    "google.cloud.bigquery",
    "googleapiclient.discovery",
    "googleapiclient.http"
]
ignore_missing_imports = true

[tool.codespell]
skip = "./.venv,uv.lock,**/*.ipynb"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pipeline"]
